added footer and instructions for user

+++ src/app/drawings/drawings.component.ts
@@ -11,19 +11,31 @@ export class DrawingsComponent implements OnInit {
 
   sheets: Sheet[] = [];
   isPrintShown = false;
+  choosenProject: string;
 
   constructor(private sheetService: SheetService) { }
 
+  checkIfSheetsExist(): boolean {
+    if(this.sheets.length === 0) {
+      return false;
+    }
+    return true;
+  }
+
+  getProjectIfChoosen() {
+    this.choosenProject = localStorage.getItem('project');
+}
+
 
   getSheets(): void {
-    // if (this.project) {
+    if (this.choosenProject) {
       this.sheetService.getSheets()
         .subscribe(sheets => this.sheets = sheets);
-    // }
+    }
   }
 
   addRow(): void {
-    // if (this.project) {
+    if (this.choosenProject) {
       const newSheet: Sheet = {
         id: null,
         no: '',
@@ -34,11 +46,13 @@ export class DrawingsComponent implements OnInit {
       };
       this.sheetService.addSheet(newSheet)
         .subscribe(sheet => this.sheets.push(sheet));
-    // }
+    } else {
+      alert('First select the project for which you want to create a new drawing.');
+    }
   }
 
   addSheet(sheet: Sheet): void {
-    // if (this.project) {
+    if (this.choosenProject) {
     const newSheet: Sheet = {
       id: sheet.id,
       no: sheet.no,
@@ -49,14 +63,16 @@ export class DrawingsComponent implements OnInit {
     };
     this.sheetService.updateSheet(newSheet)
       .subscribe();
-    // }
+    }
+
   }
 
   deleteSheet(sheet: Sheet): void {
-    // if (this.project && confirm('Are you sure you want to delete this sheet?')) {
+    if (this.choosenProject && confirm('Are you sure you want to delete this sheet?')) {
       this.sheets = this.sheets.filter(s => s !== sheet);
       this.sheetService.deleteSheet(sheet).subscribe();
-    // }
+    }
+
   }
 
   printSheets(): void {
@@ -81,7 +97,8 @@ export class DrawingsComponent implements OnInit {
 
 
   ngOnInit(): void {
-    this.getSheets();
+      this.getSheets();
+    this.getProjectIfChoosen();
   }
 
 }
